// frontend-v2/src/app/core/api/models/categoria-output.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CategoriaDisponibilidadeDto } from '../models/categoria-disponibilidade-dto';
import { CategoriaOpcaoOutput } from '../models/categoria-opcao-output';

/**
 * Dados de saida da categoria
 */
export interface CategoriaOutput {

  /**
   * Indica se a categoria esta ativa (soft delete)
   */
  ativo: boolean;

  /**
   * Descricao da categoria
   */
  descricao?: string;

  /**
   * Dias da semana em que a categoria esta disponivel
   */
  disponivel: CategoriaDisponibilidadeDto;

  /**
   * Horario final (hh:mm)
   */
  fim?: string;

  /**
   * ID da categoria
   */
  id_categoria: number;

  /**
   * ID da culinaria da categoria
   */
  id_culinaria: number;

  /**
   * Horario inicial (hh:mm)
   */
  inicio?: string;

  /**
   * Nome da categoria
   */
  nome: string;

  /**
   * Usada em pizzas: '', 'M' (valor medio) ou 'V' (maior valor)
   */
  opcao_meia?: string;

  /**
   * Opcoes desta categoria
   */
  opcoes?: Array<CategoriaOpcaoOutput>;

  /**
   * Ordem da categoria no cardapio
   */
  ordem?: number;
}